Backend/requirements.txt
Flask
Flask-Cors
fsspec
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
gunicorn
huggingface-hub
jsonpointer
jsonschema
jsonschema-specifications
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
keras
Markdown
requests
requests-oauthlib
rfc3339-validator
rfc3986-validator
transformers==4.36.2
virtualenv==20.26.6
openai-whisper==20231117  
ffmpeg==1.4  

